refactor: 그룹 상세 페이지 헤더 컴포넌트 분리
