+++ src/components/groupDetail/GroupDetailHeader.tsx
@@ -1,14 +1,48 @@
+'use client';
+
 import URLS from '@/constants/urls';
 import { IconLogo } from '@/lib/icon/Icon_ Logo';
+import { GroupDetailMode } from '@/types/groupTypes';
 import { Link } from 'lucide-react';
 
-const GroupDetailHeader = ({
-  groupTitle,
-  handleChangeMode,
-}: {
+type Props = {
   groupTitle: string;
-  handleChangeMode: () => JSX.Element;
-}) => {
+  mode: GroupDetailMode;
+  setMode: React.Dispatch<React.SetStateAction<GroupDetailMode>>;
+};
+
+const GroupDetailHeader = ({ groupTitle, mode, setMode }: Props) => {
+  /** 조건부 렌더링 버튼 */
+  const handleChangeMode = () => {
+    switch (mode) {
+      case GroupDetailMode.map:
+        return (
+          <button onClick={() => setMode(GroupDetailMode.album)}>
+            <img src='/svgs/Swap_Btn_To_Album.svg' />
+          </button>
+        );
+      case GroupDetailMode.album:
+        return (
+          <button onClick={() => setMode(GroupDetailMode.map)}>
+            <img src='/svgs/Swap_Btn_To_Map.svg' />
+          </button>
+        );
+      case GroupDetailMode.member:
+        return (
+          <button
+            className='h-10 w-10'
+            onClick={() => setMode(GroupDetailMode.album)}
+          >
+            <img
+              className='mx-auto'
+              src='/svgs/Close_Member_List.svg'
+            />
+          </button>
+        );
+      default:
+        throw new Error('잘못된 요청입니다.');
+    }
+  };
   return (
     <header className='z-50 flex h-[56px] items-center justify-between bg-white px-4 py-2'>
       <Link href={URLS.home}>
